requirements.txt
alembic==1.12.0
amqp==5.2.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
autoflake==2.2.1
bcrypt==4.0.1
billiard==4.2.0
black==23.11.0
celery==5.3.5
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.7
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
pyasn1==0.5.0
pycodestyle==2.11.1
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
pyright==1.1.336
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
sentry-sdk==1.35.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.2
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.17.0
vine==5.1.0
watchfiles==0.20.0
wcwidth==0.2.9
websockets==11.0.3
WTForms==3.0.1
